evan feedback part 2 - refactoring into smaller chunks
