+++ tests/unit/cartography/intel/okta/test_admin_roles.py
@@ -1,5 +1,5 @@
-from cartography.intel.okta.oktaintel import transform_group_roles_data
-from cartography.intel.okta.oktaintel import transform_user_roles_data
+from cartography.intel.okta.roles import transform_group_roles_data
+from cartography.intel.okta.roles import transform_user_roles_data
 from tests.data.okta.adminroles import LIST_ASSIGNED_GROUP_ROLE_RESPONSE
 from tests.data.okta.adminroles import LIST_ASSIGNED_USER_ROLE_RESPONSE
