+++ tests/unit/cartography/intel/okta/test_application.py
@@ -1,6 +1,6 @@
-from cartography.intel.okta.oktaintel import transform_application_users
-from cartography.intel.okta.oktaintel import transform_application_assigned_groups
-from cartography.intel.okta.oktaintel import transform_okta_application
+from cartography.intel.okta.applications import transform_application_users
+from cartography.intel.okta.applications import transform_application_assigned_groups
+from cartography.intel.okta.applications import transform_okta_application
 from tests.data.okta.application import create_test_application
 from tests.data.okta.application import LIST_APPLICATION_GROUP_ASSIGNED_RESPONSE
 from tests.data.okta.application import LIST_APPLICATION_USER_ASSIGNED_RESPONSE
